requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
python-jose
passlib[bcrypt]
python-dotenv